refactor(mobile): share market insight gauge between indices
Extract the Fear & Greed gauge geometry/animation into a generic
MarketInsightGauge component in mvvm/components, used by both the Mood
index (Fear & Greed) and Season (Altcoin Season) cards. The Altcoin
gauge previously had its own divergent geometry; it now renders the
identical gauge, differing only by gradient colors. Also aligns the
Season card's pressed state with the Mood index card.

LIVE-32022
