+++ .changeset/share-market-insight-gauge.md
@@ -1,0 +1,5 @@
+---
+"live-mobile": patch
+---
+
+Share a single `MarketInsightGauge` component between the Market screen's Mood index (Fear & Greed) and Season (Altcoin Season) cards so both render an identical gauge — same geometry, animation, sizing, and cursor — differing only by gradient colors. Aligns the Season card's pressed state with the Mood index card.
